move util and mypy_types into internal module
