+++ test_vector_generator/src/awses_test_vectors/internal/util.py
@@ -20,7 +20,7 @@ from aws_encryption_sdk.identifiers import AlgorithmSuite
 
 try:  # Python 3.5.0 and 3.5.1 have incompatible typing modules
     from typing import Any, Callable, Iterable, Type, Union  # noqa pylint: disable=unused-import
-    from awses_test_vectors.mypy_types import MANIFEST_VERSION  # noqa pylint: disable=unused-import
+    from awses_test_vectors.internal.mypy_types import MANIFEST_VERSION  # noqa pylint: disable=unused-import
 except ImportError:  # pragma: no cover
     # We only actually need these imports when running the mypy checks
     pass
