+++ test_vector_generator/src/awses_test_vectors/manifests/full_message/decrypt.py
@@ -23,11 +23,11 @@ import aws_encryption_sdk
 import six
 
 from awses_test_vectors.manifests.master_key import MasterKeySpec
-from awses_test_vectors.util import dictionary_validator, iterable_validator, validate_manifest_type
+from awses_test_vectors.internal.util import dictionary_validator, iterable_validator, validate_manifest_type
 
 try:  # Python 3.5.0 and 3.5.1 have incompatible typing modules
     from typing import Any, Callable, Dict, IO  # noqa pylint: disable=unused-import
-    from awses_test_vectors.mypy_types import (  # noqa pylint: disable=unused-import
+    from awses_test_vectors.internal.mypy_types import (  # noqa pylint: disable=unused-import
         DECRYPT_SCENARIO_SPEC,
         FULL_MESSAGE_DECRYPT_MANIFEST,
     )
