+++ test_vector_generator/src/awses_test_vectors/manifests/keys.py
@@ -20,7 +20,7 @@ import base64
 import attr
 import six
 
-from awses_test_vectors.util import (
+from awses_test_vectors.internal.util import (
     dictionary_validator,
     iterable_validator,
     membership_validator,
@@ -28,7 +28,7 @@ from awses_test_vectors.util import (
 )
 
 try:  # Python 3.5.0 and 3.5.1 have incompatible typing modules
-    from awses_test_vectors.mypy_types import (  # noqa pylint: disable=unused-import
+    from awses_test_vectors.internal.mypy_types import (  # noqa pylint: disable=unused-import
         AWS_KMS_KEY_SPEC,
         MANUAL_KEY_SPEC,
         KEY_SPEC,
