+++ test_vector_generator/src/awses_test_vectors/manifests/master_key.py
@@ -23,10 +23,10 @@ from aws_encryption_sdk.key_providers.kms import KMSMasterKeyProvider
 from aws_encryption_sdk.key_providers.raw import RawMasterKey
 
 from awses_test_vectors.manifests.keys import KeysManifest, KeySpec  # noqa pylint: disable=unused-import
-from awses_test_vectors.util import membership_validator
+from awses_test_vectors.internal.util import membership_validator
 
 try:  # Python 3.5.0 and 3.5.1 have incompatible typing modules
-    from awses_test_vectors.mypy_types import MASTER_KEY_SPEC  # noqa pylint: disable=unused-import
+    from awses_test_vectors.internal.mypy_types import MASTER_KEY_SPEC  # noqa pylint: disable=unused-import
 except ImportError:  # pragma: no cover
     # We only actually need these imports when running the mypy checks
     pass
